removed one circular dependence
