+++ src/asgard_dimension_tests.cpp
@@ -1,0 +1,6 @@
+#include "asgard_dimension.hpp"
+
+int main(int, char *[])
+{
+  return 0;
+}
